feat: login handler for executive register club
