+++ packages/web/src/app/executive/register-club/page.tsx
@@ -2,69 +2,35 @@
 
 import React, { useEffect, useState } from "react";
 
-import styled from "styled-components";
-
 import Custom404 from "@sparcs-clubs/web/app/not-found";
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
-import ExecutiveRegistrationTable from "@sparcs-clubs/web/common/components/ExecutiveRegistrationTable";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import PageHead from "@sparcs-clubs/web/common/components/PageHead";
-import Pagination from "@sparcs-clubs/web/common/components/Pagination";
-import { RegisterClubList } from "@sparcs-clubs/web/features/executive/register-club/services/_mock/mockRegisterClub";
-import { useGetRegisterClub } from "@sparcs-clubs/web/features/executive/register-club/services/useGetRegisterClub";
-import { useGetProfileNow } from "@sparcs-clubs/web/hooks/getProfileNow";
+import LoginRequired from "@sparcs-clubs/web/common/frames/LoginRequired";
+import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
+import { ExecutiveRegistrationClubFrame } from "@sparcs-clubs/web/features/executive/register-club/frames/ExecutiveRegistrationClubFrame";
 
 const ExecutiveRegisterClub = () => {
-  const profile = useGetProfileNow();
-
-  const [currentPage, setCurrentPage] = useState(1);
-  const limit = 10;
-
-  const { data, isLoading, isError } = useGetRegisterClub({
-    pageOffset: 1,
-    itemCount: 10,
-  });
-  const [clubData, setClubData] = useState<RegisterClubList>({
-    items: [],
-    total: 0,
-    offset: 0,
-  });
-  const [paginatedData, setPaginatedData] = useState<RegisterClubList>({
-    items: [],
-    total: 0,
-    offset: 0,
-  });
-  console.log(clubData.items.length);
+  const { isLoggedIn, login, profile } = useAuth();
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    if (!isLoading && data) {
-      setClubData(data);
-      setPaginatedData({
-        total: data?.total,
-        items: data?.items.slice(
-          (currentPage - 1) * limit,
-          currentPage * limit,
-        ),
-        offset: (currentPage - 1) * limit,
-      });
+    if (isLoggedIn !== undefined || profile !== undefined) {
+      setLoading(false);
     }
-  }, [isLoading, data, currentPage]);
+  }, [isLoggedIn, profile]);
 
-  const handlePageChange = (page: number) => {
-    setCurrentPage(page);
-  };
+  if (loading) {
+    return <AsyncBoundary isLoading={loading} isError />;
+  }
+
+  if (!isLoggedIn) {
+    return <LoginRequired login={login} />;
+  }
 
   if (profile !== "executive") {
     return <Custom404 />;
   }
-
-  const TableWithPaginationWrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 20px;
-    align-self: stretch;
-  `;
 
   return (
     <FlexWrapper direction="column" gap={20}>
@@ -75,21 +41,7 @@ const ExecutiveRegisterClub = () => {
         ]}
         title="동아리 등록 신청 내역"
       />
-      <AsyncBoundary isLoading={isLoading} isError={isError}>
-        {clubData.items.length > 0 && (
-          <TableWithPaginationWrapper>
-            <ExecutiveRegistrationTable registerList={paginatedData} />
-            <FlexWrapper direction="row" gap={16} justify="center">
-              <Pagination
-                totalPage={Math.ceil(clubData.total / limit)}
-                currentPage={currentPage}
-                limit={limit}
-                setPage={handlePageChange}
-              />
-            </FlexWrapper>
-          </TableWithPaginationWrapper>
-        )}
-      </AsyncBoundary>
+      <ExecutiveRegistrationClubFrame />
     </FlexWrapper>
   );
 };
